pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Mikhail_vlaskin <vlaskin.law@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.5"
fastapi = {extras = ["all"], version = "^0.112.1"}
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
asyncpg = "^0.29.0"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
